primera entrega avances

+++ primeraEntrega-Marrero/src/routes/products.routes.js
@@ -1,11 +1,13 @@
-import { productService }  from "../services/products.service.js";
+import { productService } from "../services/products.service.js";
+import { validate as uuidValidate } from "uuid";
 
-const productsRouter = express.Router();
+import { Router } from "express";
+
+export const productsRouter = Router();
 
 productsRouter.get("/", async (req, res) => {
     const products = await productService.getProducts();
     res.status(200).json(products);
-
 });
 
 productsRouter.get("/:pid", async (req, res) => {
@@ -17,22 +19,27 @@ productsRouter.get("/:pid", async (req, res) => {
         return res.status(404).json({ message: "Debe ingresar un id de producto" });
     }
 
+    if (!uuidValidate(pid)) {
+        return res.status(400).json({ message: "El id ingresado tiene un formato invalido" });
+    }
+
     if(!product) {
-        return res.status(404).json({ message: "Producto no encontrado" });
+        return res.status(404).json({ error: `Producto con id: ${req.params.pid} no encontrado` });
     }
 
     res.status(200).json(product);
 });
 
 productsRouter.post("/", async (req, res) => {
-    const { title, description, price, thumbnail, code, stock } = req.body;
+    const { title, description, code, price, status, stock, category, thumbnail } = req.body;
 
-    if (!title || !description || !price || !code || !stock) {
+    if (!title || !description || !code || !price || !stock || !category) {
         return res.status(400).json({ message: "Todos los campos son obligatorios, excepto thumbnail" });
     }
 
     try{
-        const product = await productService.createProduct({ title, description, price, thumbnail, code, stock });
+        const product = await productService.createProduct({ title, description, code, price, status, stock, category, thumbnail});
+        
         res.status(201).json(product);
 
     } catch(error) {
@@ -44,23 +51,39 @@ productsRouter.post("/", async (req, res) => {
 
 productsRouter.put("/:pid", async (req, res) => {
     const { pid } = req.params;
-    const { title, description, price, thumbnail, code, stock } = req.body;
+    const { title, description, code, price, status, stock, category, thumbnail } = req.body;
+
+    console.log("Datos recibidos:", pid, title, description, code, price, status, stock, category, thumbnail);
+
     if (!pid) {
         return res.status(404).json({ message: "Debe ingresar un id de producto" });
     }
 
+    if (!uuidValidate(pid)) {
+        return res.status(400).json({ message: "El id ingresado tiene un formato invalido" });
+    }
+
     if (!title || !description || !price || !code || !stock) {
-        return res.status(400).json({ message: "Todos los campos son obligatorios, excepto thumbnail" });   
+        return res.status(400).json({ message: "Todos los campos son obligatorios, excepto thumbnail" });
     }
 
-    try{
-        const product = await productService.updateProduct(pid, { title, description, price, thumbnail, code, stock });
-        res.status(200).json(product);
+    try {
+        const product = await productService.updateProduct({ id: pid, title, description, code, price, status, stock, category, thumbnail });
 
-    } catch(error) {
-        res.status(500).json({ message: "Error al actualizar el producto" });
+        if (!product) {
+            return res.status(404).json({ error: `Producto con id: ${req.params.pid} no encontrado` });
+        }
+
+        res.status(200).json({
+            message: `El producto con id: ${pid} fue actualizado correctamente.`
+        });
+
+    } catch (error)  {
+        if (error.message.includes("404")) {
+            return res.status(404).json({ error: `Product with ID: ${req.params.pid} not found` });
+          }
+          return res.status(500).json({ errors: error.message });
     }
-
 });
 
 productsRouter.delete("/:pid", async (req, res) => {
@@ -70,11 +93,23 @@ productsRouter.delete("/:pid", async (req, res) => {
         return res.status(404).json({ message: "Debe ingresar un id de producto" });
     }
 
+    if (!uuidValidate(pid)) {
+        return res.status(400).json({ message: "El id ingresado tiene un formato invalido" });
+    }
+
     try{
         const product = await productService.deleteProduct(pid);
-        res.status(200).json(product);
+        if (!product)
+            return res.status(404).json({ error: `Producto con id: ${req.params.pid} no encontrado` });
+
+        res.status(200).json({
+            message: `El producto con id: ${pid} fue eliminado correctamente.`,
+            product: product
+          });
 
     } catch(error) {
-        res.status(500).json({ message: "Error al eliminar el producto" });
+          return res.status(500).json({ errors: error.message });
     }
 });
+
+
